refactor : PurchaseInfo 생성 및 Lottos, LottoSimulator 리팩터링
로또 가격, 구매 로또 장수 등 구매와 관련된 PurchaseInfo 생성
Lottos 및 LottoSimulator 리팩터링
Lottos의 checkResult indent 제약 조건을 위반한 부분 리팩터링

+++ src/main/java/view/LottoSimulatorView.java
@@ -14,9 +14,9 @@ public class LottoSimulatorView {
     public static final String ASK_LOTTO_NUMBER_FOR_ANSWER_LOTTO_MESSAGE = "지난 주 당첨 번호를 입력해주세요.";
     public static final String ASK_BONUS_NUMBER_FOR_ANSWER_LOTTO_MESSAGE = "보너스 볼을 입력해주세요.";
 
-    public int askMoneyForBuyLotto() {
+    public long askMoneyForBuyLotto() {
         System.out.println(ASK_MONEY_MESSAGE);
-        return Integer.parseInt(scanner.nextLine());
+        return Long.parseLong(scanner.nextLine());
     }
 
     public String askLottoNumberForAnswerLotto() {
